requirements.txt
msrest==0.7.1
azure-cognitiveservices-vision-face==0.6.0
numpy==1.24.2
imutils==0.5.4
Pillow==9.5.0
pyimgur==0.6.0
python-telegram-bot==20.2
opencv-python==4.7.0.72
cython==0.29.24
setuptools-rust==1.5.2

